+++ src/anyio/_backends/_trio.py
@@ -10,6 +10,7 @@ from os import PathLike
 from signal import Signals
 from types import TracebackType
 from typing import (
+    IO,
     Any,
     Awaitable,
     Callable,
@@ -380,9 +381,9 @@ async def open_process(
     command: Union[str, bytes, Sequence[Union[str, bytes]]],
     *,
     shell: bool,
-    stdin: int,
-    stdout: int,
-    stderr: int,
+    stdin: Union[int, IO[Any], None],
+    stdout: Union[int, IO[Any], None],
+    stderr: Union[int, IO[Any], None],
     cwd: Union[str, bytes, PathLike, None] = None,
     env: Optional[Mapping[str, str]] = None,
     start_new_session: bool = False,