+++ docs/versionhistory.rst
@@ -6,7 +6,11 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 **UNRELEASED**
 
 - Fixed ``TypeError`` in ``get_current_task()`` on asyncio when using a custom ``Task`` factory
-- Updated type annotations on ``open_process()`` to accept bytes and sequences of bytes
+- Updated type annotations on ``run_process()`` and ``open_process()``:
+
+  * ``command`` now accepts accepts bytes and sequences of bytes
+  * ``stdin``, ``stdout`` and ``stderr`` now accept file-like objects
+    (PR by John T. Wodder II)
 
 **3.5.0**